+++ docs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.6.1 (2022-02-11)
 
-- fix: assertion function and json number parse rule
+- fix: assertion function errors and json number parse rule
 
 ## v0.6.0 (2022-02-08)